chore(ui): reorganize window files into chrome directory
